Load token/profile from CLI output (rather than file)
As Az CLI is gonna deprecate accessTokens.json and azureProfile.json in
the near future, we shall not loading them from file.

Az CLI comes with commands which has same effect as parsing files:

- `az account show [-s subid]`: returns a json object representing the
  specified subscription object (the default one if `subid` not
  specified)
- `az account get-access-token`: returns a json object representing the
  specified token object.

Current implementation, when building for azure cli auth method, it
loads the *azureProfile.json* file and parse it into a `profile` object.
Then it will first resolve the`subscriptionId` and populate the `clientId`,
`tenantId`, `environment` up to the `azureCLIProfile`.

Especially, when populating `clientId`, it will load *accessTokens.json*
file and iterates all the token objects contained in it to find the "arm"
token with correct `tenantId`, use its `clientId` to populate back.

However, `clientID` should be used together with `refreshToken` in order to
refresh the `accessToken`. As the output of `az account get-access-token` will
not output both `clientID` and `refreshToken`, there is no way to get
them. On the other hand, inside go-autorest, it will invoke the refresh
token logic only when there is no custom refresh function set. In our
case we are setting the custom refresh logic to call `az cli
get-access-token` to refresh the new access token. This means, there is
no need to set `clientID`/`refreshToken` in our case.
(go-autorest will check empty for `clientID` though, so I just pass in a
dummy clientID to make the validation pass).

Another change is that rather than reading from *azureProfile.json* to
get all the subscriptions, we can just read the one user specified
in provider block/env var via `az account show -s [subid]` . Or loading
the default one if user didnt explicitly specified that.