experimental: poc for a dynamic recaser
This approach requires a minor change to each Resource ID type (which actually reduces the TLOC) - but equally
would only work where a given Resource ID Type has been imported. This approach would mean that importing anything
from `commonids` would allow recasing of those IDs, but less common Resource ID types (for example Automation Configurations)
would only be recased if used in the SDK - which is likely fine within the AzureRM Provider (since we import basically everything)
however could become problematic when importing only a subset.

This means that when used as so:

```
package main

import (
	"log"

	"github.com/hashicorp/go-azure-helpers/resourcemanager/commonids"
	"github.com/hashicorp/go-azure-helpers/resourcemanager/resourceids/recaser"
)

func main() {
	// this is present just to ensure the package gets registered
	resourceGroupId := commonids.NewResourceGroupID("0000", "hello")
	log.Printf(resourceGroupId.ID())

	log.Printf(recaser.ReCase("/subscriptions/11111/resourcegroups/bobby/providers/Microsoft.Compute/availabilitySets/HeYO"))
}
```

The following is output:

```
$ go mod vendor && go build . && ./example-helpers
2023/10/16 13:33:22 /subscriptions/0000/resourceGroups/hello
2023/10/16 13:33:22 /subscriptions/11111/resourceGroups/bobby/providers/Microsoft.Compute/availabilitySets/HeYO
```

However if the Resource ID type isn't referenced (in this case, from the `commonids` package), then the original value remains:

```
 $ go mod vendor && go build . && ./example-helpers
2023/10/16 14:15:50 /subscriptions/11111/resourcegroups/bobby/providers/Microsoft.Compute/availabilitySets/HeYO
```

We likely need to special-case a handful of keys (`subscriptions`, `resourceGroups`, `providers` etc) in addition to this
but that should be a subset.