`resourcemanager/resourceids`: adding a `Match` function
The `Match` function takes two instances of the ResourceId type and allows comparing the parsed segments for these.

Whilst this is also possible by comparing the value of `.ID()` in both instances - that's not case-aware and so will be problematic
in the future - hence the need for this function which can account for case-aware comparisons.

This also exposed a feature-flag for treating User Specified Segments as case-insensitive for comparison purposes - however
this isn't usable at this time and so MUST NOT be exposed as a toggle at this point-in-time, since it'll cause more problems
than it solves. However having this internal-only feature flag allows us to built out the surrounding components ahead-of-time
and add tests covering both scenarios - therefore this is beneficial to have as an internal toggle today - even if it's not
usable for end-users for some time.